src/requirements.txt
pandas
PyPDF2
sqlalchemy
psycopg2-binary
python-dotenv
xlsxwriter
requests
boto3
openpyxl